create script work with existing projects

+++ create.js
@@ -12,33 +12,43 @@ console.log(`Creating project: ${config.id}`)
 
 globalThis.projectPath = path.join("projects", config.id)
 
+let newProject
 if (fs.existsSync(projectPath)) {
   globalThis.project = JSON.parse(fs.readFileSync(projectPath + "/project.json"))
-  project.config = structuredClone(config)
+  globalThis.config = project.config
 } else {
+  newProject = true
   fs.mkdirSync(projectPath)
   globalThis.project = {
-    config: structuredClone(config),
+    config: config,
     curseforge: {},
     planetminecraft: {},
     modrinth: {}
   }
+
+  fs.cpSync("data/create/images", path.join(projectPath, "images"), { recursive: true })
+  fs.cpSync("data/create/pack.png", path.join(projectPath, "pack.png"))
+
+  if (fs.existsSync("data/create/thumbnail.png")) {
+    fs.cpSync("data/create/thumbnail.png", path.join(projectPath, "thumbnail.png"))
+  }
+
+  if (fs.existsSync("data/create/logo.png")) {
+    fs.cpSync("data/create/logo.png", path.join(projectPath, "logo.png"))
+  }
 }
 save()
 
-// Copying
-
 const templates = path.join(projectPath, "templates")
-fs.mkdirSync(templates)
-for (const entry of fs.readdirSync("templates")) {
-  if (entry === "snippets") continue
-  const src = path.join("templates", entry)
-  const dest = path.join(templates, entry)
-  fs.cpSync(src, dest)
+if (!fs.existsSync(templates)) {
+  fs.mkdirSync(templates)
+  for (const entry of fs.readdirSync("templates")) {
+    if (entry === "snippets") continue
+    const src = path.join("templates", entry)
+    const dest = path.join(templates, entry)
+    fs.cpSync(src, dest)
+  }
 }
-
-fs.cpSync("data/create/images", path.join(projectPath, "images"), { recursive: true })
-fs.cpSync("data/create/pack.png", path.join(projectPath, "pack.png"))
 
 // Files
 
@@ -50,27 +60,25 @@ for (const img of config.images) {
   img.buffer = await sharp(path.join(projectPath, "images", img.file + ".png")).resize(1920, 1080, { fit: "inside", withoutEnlargement: true }).jpeg({ quality: 95 }).toBuffer()
 }
 
-if (fs.existsSync("data/create/thumbnail.png")) {
-  fs.cpSync("data/create/thumbnail.png", path.join(projectPath, "thumbnail.png"))
+const thumbnailPath = path.join(projectPath, "thumbnail.png")
+if (fs.existsSync(thumbnailPath)) {
   config.images.unshift({
     name: "Project Thumbnail",
     description: `The thumbnail image for ${config.name}`,
     file: "thumbnail",
     thumbnail: true,
-    buffer: await sharp("data/create/thumbnail.png").resize(1920, 1080, { fit: "inside", withoutEnlargement: true }).jpeg({ quality: 95 }).toBuffer()
+    buffer: await sharp(thumbnailPath).resize(1920, 1080, { fit: "inside", withoutEnlargement: true }).jpeg({ quality: 95 }).toBuffer()
   })
 }
 
-if (fs.existsSync("data/create/logo.png")) {
-  if (config.logo) {
-    fs.cpSync("data/create/logo.png", path.join(projectPath, "logo.png"))
-  }
+const logoPath = path.join(projectPath, "logo.png")
+if (fs.existsSync(logoPath)) {
   config.images.push({
     name: "Project Logo",
     description: `The logo for ${config.name}`,
     file: "logo",
     logo: true,
-    buffer: await sharp("data/create/logo.png").resize(1280, 256, { fit: "inside", withoutEnlargement: true }).toBuffer()
+    buffer: await sharp(logoPath).resize(1280, 256, { fit: "inside", withoutEnlargement: true }).toBuffer()
   })
 }
 
@@ -106,7 +114,7 @@ if (!project.modrinth.id) {
 
 // Ewan Howell
 
-if (settings.ewan) {
+if (settings.ewan && newProject) {
   await ewanhowell.writeDetails()
   await ewanhowell.addImages()
